fix private revisions
If you went to your own page and you had a 100% private
piece then if you update the piece it would show the
1st revision and there'd be no revisions icon.

That's fixed although the UX here is kind of broken
Your page shows either the most recent PUBLIC revision
or the newest private revision. Maybe it should always
show the newest revision if you're logged in?