add UI anim
what a PITA!!!

This feels so wrong but I just wanted it to be done. Basically I wanted
to try to make it clear through animation what the UI button is so I make
the 2 panes fly into the button.

I wanted to do that with CSS animations but it's a point in the ass
because CSS animations are not designed to be done by JavaScript
so I'd have to look up the positions, compute a CSS animation,
then set it off. Then I have to trigger hidding the elements
after it finishes.

I looked into it for 10-20 mins and decided fuck it, I'm just
going to draw my own rect. People will get the point. I can revisit
later maybe never.